Moved universal php files into /require, replaced theme button by icon
