+++ post.php
@@ -1,17 +1,14 @@
 <!doctype html>
 <html>
 <head>
-<head>
-  <?php require('head.php');?>
+  <?php require('require/head.php');?>
 </head>
 <body>
-  <div id="header">
-    <a href="index.php" title="jaseph.com"><img id="headerimg" src="assets/jaseph_black.png"/></a>
-  </div>
+  <?php require('require/header.php');?>
   <div class="content">
     <?php
     // Credentials for this server
-    require('credentials.php');
+    require('require/credentials.php');
 
     $success = true;
 
@@ -53,7 +50,6 @@
   }
     ?>
     <br>
-    <button id="swapper" onclick="swapStyle()">Hacker Mode</button>
   </div>
   <div id="footer">
     <a href="https://github.com/phmrz/JasephCMS" title="Check out the main branch of this page on GitHub!">Created as a school project by Jakob Mainka, Philipp Merz and Sebastian Scheinert</a>
