+++ require/header.php
@@ -1,0 +1,5 @@
+
+  <div id="header">
+    <a href="../JasephCMS" title="jaseph"><img id="headerimg" src="assets/jaseph_black.png" alt=""/></a>
+    <img id="themeselector" src="assets/themes.png" onclick="swapStyle()" title="Change the page theme!" alt=""/>
+  </div>
